refactor(authentication,chat,database,forms,push-notifications,storage): application route files
