+++ modules/push-notifications/src/PushNotifications.ts
@@ -6,7 +6,7 @@ import {
 } from '@conduitplatform/grpc-sdk';
 import AppConfigSchema, { Config } from './config';
 import { AdminHandlers } from './admin';
-import { PushNotificationsRoutes } from './routes/routes';
+import { PushNotificationsRoutes } from './routes';
 import * as models from './models';
 import {
   GetNotificationTokensRequest,
